pyproject.toml
[project]
name = "ai-native-base"
version = "0.1.0"
description = "AI-native product development workflow template for Codex-driven Ralph loops."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.11,<3",
  "PyYAML>=6.0,<7"
]

[project.scripts]
ainative = "ai_native.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.4,<9",
  "pytest-cov>=6.0,<7",
  "ruff>=0.11,<0.12"
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai_native"]
